Changed GlobalLocationNumber to long since it won't fit in an int
#246

+++ Models/AttentionData/GlobalLocationNumber.cs
@@ -2,7 +2,7 @@
 {
     public class GlobalLocationNumber
     {
-        public int globalLocationNumber { get; set; }
+        public long globalLocationNumber { get; set; }
 
         public string location { get; set; }
     }
